StateMachine: move content to plutus-contract
