+++ plutus-contract/plutus-contract.cabal
@@ -54,6 +54,7 @@ library
         Language.Plutus.Contract.Servant
         Language.Plutus.Contract.Resumable
         Language.Plutus.Contract.StateMachine
+        Language.Plutus.Contract.StateMachine.OnChain
         Language.Plutus.Contract.Tx
         Language.Plutus.Contract.Util
         Language.Plutus.Contract.Wallet
