+++ plutus-contract/src/Language/Plutus/Contract/StateMachine/OnChain.hs
@@ -12,7 +12,7 @@
 --   define a @StateMachine s i@ with input type @i@ and state type @s@. Then
 --   use 'mkValidator' in on-chain code to check the required hashes and
 --   validate the transition, and 'mkRedeemer' to make redeemer scripts.
-module Language.PlutusTx.StateMachine(
+module Language.Plutus.Contract.StateMachine.OnChain(
       StateMachine(..)
     , StateMachineInstance (..)
     , State(..)
